# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateReview:
    properties:
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      content_type:
        enum:
        - anime
        - game
        - movie
        - tv
        type: string
      review:
        maxLength: 1000
        minLength: 6
        type: string
      star:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - content_id
    - content_type
    - review
    - star
    type: object
...
  requests.ID:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  requests.SortReviewByUserID:
    properties:
      page:
        minimum: 1
        type: integer
      sort:
        enum:
        - popularity
        - latest
        - oldest
        type: string
      userID:
        type: string
    required:
    - page
    - sort
    - userID
    type: object
...
  responses.Author:
    properties:
      _id:
        type: string
      email:
        type: string
      image:
        type: string
      is_premium:
        type: boolean
      username:
        type: string
    type: object
...
  responses.Review:
    properties:
      _id:
        type: string
      author:
        $ref: '#/definitions/responses.Author'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      created_at:
        type: string
      is_author:
        type: boolean
      is_liked:
        type: boolean
      likes:
        items:
          type: string
        type: array
      popularity:
        type: integer
      review:
        type: string
      star:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: http://localhost:8080
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Project Consumer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: '#soon'
  title: Project Consumer API
  version: "1.0"
paths:
...
  /review:
    delete:
      consumes:
      - application/json
      description: Deletes Review
      parameters:
      - description: ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Review
      tags:
      - review
    get:
      consumes:
      - application/json
      description: Get Reviews by User
      parameters:
      - description: Sort Review by User ID
        in: body
        name: sortreviewbyuserid
        required: true
        schema:
          $ref: '#/definitions/requests.SortReviewByUserID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.Review'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Reviews by User
      tags:
      - review
    patch:
      consumes:
      - application/json
      description: Update Review
      parameters:
      - description: Create Review
        in: body
        name: createreview
        required: true
        schema:
          $ref: '#/definitions/requests.CreateReview'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Review'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Review
      tags:
      - review
    post:
      consumes:
      - application/json
      description: Creates Review
      parameters:
      - description: Create Review
        in: body
        name: createreview
        required: true
        schema:
          $ref: '#/definitions/requests.CreateReview'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.Review'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Review
      tags:
      - review
  /review/vote:
    patch:
      consumes:
      - application/json
      description: Like Review
      parameters:
      - description: ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Review'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Vote Review
      tags:
      - review
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"